requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.1.2
Pillow==10.2.0
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
openpyxl==3.1.2
reportlab==4.1.0
pandas==2.2.2
Werkzeug==3.0.1
WTForms==3.1.1
Flask-WTF==1.2.1
requests==2.32.3